requirements.txt
python-telegram-bot[webhooks]==21.0
feedparser
python-dotenv
beautifulsoup4
aiohttp
APScheduler
requests
aiohttp
APScheduler
beautifulsoup4
feedparser
python-dotenv
python-telegram-bot[webhooks]==21.0
requests
telegramutils.py
